Adjusted suggestions
