+++ src/main/java/com/mercadopago/net/Headers.java
@@ -1,14 +1,22 @@
 package com.mercadopago.net;
 
 public class Headers {
-    public static String AUTHORIZATION = "Authorization";
-    public static String CONTENT_TYPE = "Content-Type";
-    public static String ACCEPT = "Accept";
-    public static String USER_AGENT = "User-Agent";
-    public static String IDEMPOTENCY_KEY = "X-Idempotency-Key";
-    public static String PRODUCT_ID = "X-Product-Id";
-    public static String CORPORATION_ID = "X-Corporation-Id";
-    public static String INTEGRATOR_ID = "X-Integrator-Id";
-    public static String PLATFORM_ID = "X-Platform-Id";
+  public static String AUTHORIZATION = "Authorization";
+
+  public static String CONTENT_TYPE = "Content-Type";
+
+  public static String ACCEPT = "Accept";
+
+  public static String USER_AGENT = "User-Agent";
+
+  public static String IDEMPOTENCY_KEY = "X-Idempotency-Key";
+
+  public static String PRODUCT_ID = "X-Product-Id";
+
+  public static String CORPORATION_ID = "X-Corporation-Id";
+
+  public static String INTEGRATOR_ID = "X-Integrator-Id";
+
+  public static String PLATFORM_ID = "X-Platform-Id";
 
 }
