+++ src/main/java/com/mercadopago/net/KeepAliveStrategy.java
@@ -10,20 +10,21 @@ import org.apache.http.protocol.HttpContext;
 
 public class KeepAliveStrategy implements ConnectionKeepAliveStrategy {
 
-    private static final int DEFAULT_KEEP_ALIVE_TIMEOUT_MS = 10000;
-    private static final String KEEP_ALIVE_TIMEOUT_PARAM_NAME = "timeout";
+  private static final int DEFAULT_KEEP_ALIVE_TIMEOUT_MS = 10000;
 
-    @Override
-    public long getKeepAliveDuration(HttpResponse response, HttpContext context) {
-        HeaderElementIterator it = new BasicHeaderElementIterator(response.headerIterator(HTTP.CONN_KEEP_ALIVE));
-        while (it.hasNext()) {
-            HeaderElement he = it.nextElement();
-            String param = he.getName();
-            String value = he.getValue();
-            if (value != null && param.equalsIgnoreCase(KEEP_ALIVE_TIMEOUT_PARAM_NAME)) {
-                return Long.parseLong(value) * 1000;
-            }
-        }
-        return DEFAULT_KEEP_ALIVE_TIMEOUT_MS;
+  private static final String KEEP_ALIVE_TIMEOUT_PARAM_NAME = "timeout";
+
+  @Override
+  public long getKeepAliveDuration(HttpResponse response, HttpContext context) {
+    HeaderElementIterator it = new BasicHeaderElementIterator(response.headerIterator(HTTP.CONN_KEEP_ALIVE));
+    while (it.hasNext()) {
+      HeaderElement he = it.nextElement();
+      String param = he.getName();
+      String value = he.getValue();
+      if (value != null && param.equalsIgnoreCase(KEEP_ALIVE_TIMEOUT_PARAM_NAME)) {
+        return Long.parseLong(value) * 1000;
+      }
     }
+    return DEFAULT_KEEP_ALIVE_TIMEOUT_MS;
+  }
 }
